changed serving data type from integer to decimal

+++ db/migrate/20200518230631_create_nutrientprofiles.rb
@@ -2,7 +2,7 @@ class CreateNutrientprofiles < ActiveRecord::Migration[6.0]
   def change
     create_table :nutrientprofiles do |t|
       t.string :name
-      t.integer :serving
+      t.decimal :serving
       t.string :size
       t.integer :weight
       t.integer :calories, default: 0
